Move SharedBspFileOptions.bspDetails to SetupIde
