+++ modules/cli/src/main/scala/scala/cli/commands/SetupIde.scala
@@ -12,7 +12,7 @@ import scala.build.Inputs.WorkspaceOrigin
 import scala.build.errors.{BuildException, WorkspaceError}
 import scala.build.internal.{Constants, CustomCodeWrapper}
 import scala.build.options.{BuildOptions, Scope}
-import scala.build.{Artifacts, CrossSources, Inputs, Logger, Sources}
+import scala.build.{Artifacts, CrossSources, Inputs, Logger, Os, Sources}
 import scala.cli.CurrentParams
 import scala.cli.commands.util.CommonOps._
 import scala.cli.commands.util.SharedOptionsUtil._
@@ -93,7 +93,7 @@ object SetupIde extends ScalaCommand[SetupIdeOptions] {
     if (options.buildOptions.classPathOptions.extraDependencies.toSeq.nonEmpty)
       value(downloadDeps(inputs, options.buildOptions, logger))
 
-    val (bspName, bspJsonDestination) = options.bspFile.bspDetails(inputs.workspace)
+    val (bspName, bspJsonDestination) = bspDetails(inputs.workspace, options.bspFile)
     val scalaCliBspJsonDestination =
       inputs.workspace / Constants.workspaceDirName / "ide-options-v2.json"
 
@@ -146,5 +146,16 @@ object SetupIde extends ScalaCommand[SetupIdeOptions] {
     }
     else
       None
+  }
+
+  def bspDetails(workspace: os.Path, ops: SharedBspFileOptions): (String, os.Path) = {
+    import ops._
+    val dir = bspDirectory
+      .filter(_.nonEmpty)
+      .map(os.Path(_, Os.pwd))
+      .getOrElse(workspace / ".bsp")
+    val bspName0 = bspName.map(_.trim).filter(_.nonEmpty).getOrElse("scala-cli")
+
+    (bspName0, dir / s"$bspName0.json")
   }
 }
