+++ modules/cli/src/main/scala/scala/cli/commands/SharedBspFileOptions.scala
@@ -2,8 +2,6 @@ package scala.cli.commands
 
 import caseapp._
 
-import scala.build.Os
-
 // format: off
 final case class SharedBspFileOptions(
   @Name("bspDir")
@@ -14,19 +12,8 @@ final case class SharedBspFileOptions(
   @HelpMessage("Name of BSP")
   @Hidden
     bspName: Option[String] = None
-) {
-  // format: on
-
-  def bspDetails(workspace: os.Path): (String, os.Path) = {
-    val dir = bspDirectory
-      .filter(_.nonEmpty)
-      .map(os.Path(_, Os.pwd))
-      .getOrElse(workspace / ".bsp")
-    val bspName0 = bspName.map(_.trim).filter(_.nonEmpty).getOrElse("scala-cli")
-
-    (bspName0, dir / s"$bspName0.json")
-  }
-}
+)
+// format: on
 
 object SharedBspFileOptions {
   lazy val parser: Parser[SharedBspFileOptions]                           = Parser.derive
